Fixes for bugs related to SDOE in 3.21-rc (#1210)
* restricting matplotlib due to issues when using 3.8

* cast parameter for setValue to int

* Add link to CCSI-Toolset/FOQUS#1211

---------

+++ foqus_lib/gui/sdoe/sdoeAnalysisDialog.py
@@ -849,7 +849,7 @@ class sdoeAnalysisDialog(_sdoeAnalysisDialog, _sdoeAnalysisDialogUI):
                 "mwr = %d, n = %d" % (mwr, results[mwr]["num_restarts"])
             )
             count += 1
-            self.SDOE2_progressBar.setValue((100 / len(mwr_list)) * count)
+            self.SDOE2_progressBar.setValue(int((100 / len(mwr_list)) * count))
             QApplication.processEvents()
 
         self.unfreeze()
